+++ PerlinNoise.h
@@ -10,13 +10,14 @@ class CPerlinNoiseGenerator_1D
 {
 private:
 	std::vector<std::function<double(int)>*>	m_random_funcs;
-	double									m_persistence;
-	unsigned int							m_seed;
+	double										m_persistence;
+	unsigned int								m_seed;
+
 public:
 			CPerlinNoiseGenerator_1D();
 			CPerlinNoiseGenerator_1D(const CPerlinNoiseGenerator_1D& ref);
 	void	addFunction(std::function<double(int)>* func);
-	int		operator()(int) const;
+	double	operator()(int) const;
 
 	// Getter
 	double getPersistence() const	{ return m_persistence; }
@@ -38,7 +39,7 @@ public:
 	CPerlinNoiseGenerator_2D();
 	CPerlinNoiseGenerator_2D(const CPerlinNoiseGenerator_2D& ref);
 	void	addFunction(std::function<double(int,int)>* func);
-	int		operator()(int, int) const;
+	double	operator()(int, int) const;
 
 	// Getter
 	double getPersistence() const { return m_persistence; }
@@ -60,7 +61,7 @@ public:
 	CPerlinNoiseGenerator_3D();
 	CPerlinNoiseGenerator_3D(const CPerlinNoiseGenerator_3D& ref);
 	void	addFunction(std::function<double(int, int, int)>* func);
-	int		operator()(int, int, int) const;
+	double 	operator()(int, int, int) const;
 
 	// Getter
 	double getPersistence() const { return m_persistence; }
